Move utilities.utils.get_viewname() to utilities.views
