+++ netbox/extras/dashboard/widgets.py
@@ -17,7 +17,8 @@ from extras.choices import BookmarkOrderingChoices
 from netbox.choices import ButtonColorChoices
 from utilities.permissions import get_permission_for_model
 from utilities.templatetags.builtins.filters import render_markdown
-from utilities.utils import content_type_identifier, content_type_name, dict_to_querydict, get_viewname
+from utilities.utils import content_type_identifier, content_type_name, dict_to_querydict
+from utilities.views import get_viewname
 from .utils import register_widget
 
 __all__ = (
