+++ netbox/utilities/utils.py
@@ -21,39 +21,9 @@ from mptt.models import MPTTModel
 from dcim.choices import CableLengthUnitChoices, WeightUnitChoices
 from extras.utils import is_taggable
 from netbox.config import get_config
-from netbox.plugins import PluginConfig
 from utilities.constants import HTTP_REQUEST_META_SAFE_COPY
 from .constants import HTML_ALLOWED_ATTRIBUTES, HTML_ALLOWED_TAGS
 from .string import title
-
-
-def get_viewname(model, action=None, rest_api=False):
-    """
-    Return the view name for the given model and action, if valid.
-
-    :param model: The model or instance to which the view applies
-    :param action: A string indicating the desired action (if any); e.g. "add" or "list"
-    :param rest_api: A boolean indicating whether this is a REST API view
-    """
-    is_plugin = isinstance(model._meta.app_config, PluginConfig)
-    app_label = model._meta.app_label
-    model_name = model._meta.model_name
-
-    if rest_api:
-        viewname = f'{app_label}-api:{model_name}'
-        if is_plugin:
-            viewname = f'plugins-api:{viewname}'
-        if action:
-            viewname = f'{viewname}-{action}'
-
-    else:
-        viewname = f'{app_label}:{model_name}'
-        if is_plugin:
-            viewname = f'plugins:{viewname}'
-        if action:
-            viewname = f'{viewname}_{action}'
-
-    return viewname
 
 
 def csv_format(data):
